refactor: Rename fetchJobs file to fetch-jobs for consistency
