+++ frontend/src/lib/fetch-jobs.ts
